:memo: The robots were separated according to their roles.
